Fix client authentication and move volume command handlers to one file
