+++ client/volumes.go
@@ -12,11 +12,12 @@ type VolumeService struct {
 }
 
 func (s *VolumeService) Create(ctx context.Context, userID, projectID string, create types.VolumeCreateRequest) (types.Volume, error) {
-	uri := fmt.Sprintf("projects/%s/%s/volume", userID, projectID)
+	uri := fmt.Sprintf("projects/%s/%s/volumes", userID, projectID)
 	req, err := s.client.NewAuthorizedRestRequest(Post, uri, nil, create)
 	if err != nil {
 		return types.Volume{}, err
 	}
+
 	res := types.Volume{}
 	if err = s.client.ExecuteRest(ctx, req, res); err != nil {
 		return types.Volume{}, err
@@ -25,25 +26,13 @@ func (s *VolumeService) Create(ctx context.Context, userID, projectID string, cr
 	return res, nil
 }
 
-func (s *VolumeService) Update(ctx context.Context, userID, projectID, volumeIDOrName string, update types.VolumeResizeRequest) error {
-	uri := fmt.Sprintf("projects/%s/%s/volume/%s", userID, projectID, volumeIDOrName)
-	req, err := s.client.NewAuthorizedRestRequest(Put, uri, nil, update)
-	if err != nil {
-		return err
-	}
-	if err = s.client.ExecuteRest(ctx, req, nil); err != nil {
-		return err
-	}
-
-	return nil
-}
-
 func (s *VolumeService) Delete(ctx context.Context, userID, projectID, volumeIDOrName string) error {
-	uri := fmt.Sprintf("projects/%s/%s/volume/%s", userID, projectID, volumeIDOrName)
+	uri := fmt.Sprintf("projects/%s/%s/volumes/%s", userID, projectID, volumeIDOrName)
 	req, err := s.client.NewAuthorizedRestRequest(Delete, uri, nil, nil)
 	if err != nil {
 		return err
 	}
+
 	if err = s.client.ExecuteRest(ctx, req, nil); err != nil {
 		return err
 	}
@@ -52,7 +41,7 @@ func (s *VolumeService) Delete(ctx context.Context, userID, projectID, volumeIDO
 }
 
 func (s *VolumeService) List(ctx context.Context, userID string, projectID string) ([]types.Volume, error) {
-	uri := fmt.Sprintf("projects/%s/%s/volume", userID, projectID)
+	uri := fmt.Sprintf("projects/%s/%s/volumes", userID, projectID)
 	req, err := s.client.NewAuthorizedRestRequest(Get, uri, nil, nil)
 	if err != nil {
 		return nil, err
@@ -67,4 +56,18 @@ func (s *VolumeService) List(ctx context.Context, userID string, projectID strin
 	}
 
 	return res, err
+}
+
+func (s *VolumeService) Update(ctx context.Context, userID, projectID, volumeIDOrName string, update types.VolumeResizeRequest) error {
+	uri := fmt.Sprintf("projects/%s/%s/volumes/%s", userID, projectID, volumeIDOrName)
+	req, err := s.client.NewAuthorizedRestRequest(Put, uri, nil, update)
+	if err != nil {
+		return err
+	}
+
+	if err = s.client.ExecuteRest(ctx, req, nil); err != nil {
+		return err
+	}
+
+	return nil
 }
